file structure and improve graph tooltip
