+++ frontend/src/components/dashboard/dashboard-calories-graph.tsx
