+++ frontend/src/components/dashboard/dashboard-macros-graph.tsx
