+++ frontend/src/components/dashboard/loading-graph.tsx
