+++ frontend/src/components/entries/food-card.tsx
@@ -1,13 +1,13 @@
-import { FoodUSDA } from "@/lib/types"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "./ui/card"
-import { title } from "@/lib/utils"
+import { Food } from "@/lib/types"
+import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "../ui/card"
+import { title } from "@/lib/utils/utils"
 
 interface FoodCardProps {
-    food: FoodUSDA
-    onClick?: (food: FoodUSDA) => void
+    food: Food
+    onClick?: (food: Food) => void
 }
 
-export const USDAFoodCard = ({ food, onClick }: FoodCardProps) => {
+export const FoodCard = ({ food, onClick }: FoodCardProps) => {
     return (
         <Card 
             className="flex flex-col @container/card justify-between cursor-pointer hover:bg-muted hover:text-muted-foreground"
@@ -16,7 +16,6 @@ export const USDAFoodCard = ({ food, onClick }: FoodCardProps) => {
             <div className="flex flex-col gap-4">
                 <CardHeader>
                     <CardTitle>{title(food.name)}</CardTitle>
-                    <CardDescription>{!!food.brandOwner ? food.brandOwner : "Unknown Brand"}</CardDescription>
                 </CardHeader>
                 <CardContent>
                     {food.nutrients.map(nutrient => (
@@ -27,9 +26,6 @@ export const USDAFoodCard = ({ food, onClick }: FoodCardProps) => {
                     ))}
                 </CardContent>
             </div>
-            <CardFooter className="justify-end -mr-4 -mb-4">
-                <span className="text-sm font-light">FDC ID {food.id}</span>
-            </CardFooter>
         </Card>
     )
 }
