+++ frontend/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils"
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
